New: moved shared types
