--- client/src/interfaces/index.tsx
@@ -1,4 +1,0 @@
-export interface PlaygroundFeature {
-  name: string;
-  value: string;
-}
